add products route
The code contains a function that simulates a product request where the response time is from 0 to 10 seconds. In the main route, it is checked whether there is any content in cache. If there is, it is presented, otherwise the simulated function is used. Then, a new cache is created and the content is presented.