:zap: improve perfomance using rate-limit
The code takes the IP and resource, creates a key, creates a request counter based on that key, and then saves the number of requests with a 30-second time limit in REDIS. It performs a check, and if there are more than 5 requests by default during these 30 seconds, an error is issued in the request. Therefore, only one new request will be allowed after the 30 seconds have passed.